+++ a3/space.h
@@ -2,9 +2,9 @@
 #define SPACE_H
 
 typedef struct triangle {
-  int *vertex0;
-  int *vertex1;
-  int *vertex2;
+  double *vertex0;
+  double *vertex1;
+  double *vertex2;
   struct triangle *next; // triangle lists will be NULL terminated
 } Triangle;
 
@@ -15,17 +15,17 @@ typedef struct entity {
 
 // this struct is mainly for freeing vertices (fix double freeing problem if same vertex used twice)
 typedef struct vertex {
-  int *coordinate;
+  double *coordinate;
   struct vertex *next; // vertices will be NULL terminated
 } Vertex;
 
-Triangle *create_triangle(int *v1, int *v2, int *v3);
+Triangle *create_triangle(double *v1, double *v2, double *v3);
 Triangle *create_triangle_v(Vertex *v1, Vertex *v2, Vertex *v3);
 Entity *create_entity(Triangle *object);
-Vertex *create_vertex(int *coordinate);
+Vertex *create_vertex(double *coordinate);
 
 // (x,y,z) is the coordinate of the front left bottom vertex of the rectangle
-Entity *create_rectangle(Entity *entities, Vertex **vertex_list, int x, int y, int z, int x_length, int y_length, int z_length);
+Entity *create_rectangle(Entity *entities, Vertex **vertex_list, double x, double y, double z, double x_length, double y_length, double z_length);
 void free_all_triangles(Triangle *triangles);
 void free_all_entities(Entity *entities);
 void free_all_vertices(Vertex *vertices);
